Function  that prints a name.
Function that executes a function given as a parameter on each element of an array
Function that searches for an integer
Program that performs simple operations.
This file should contain all the function prototypes and data structures used by the program
This file should contain the 5 following functions
This file should contain the function that selects the correct function to perform the operation asked by the user. You’re not allowed to declare any other function
This file should contain your main function only
Program that prints the opcodes of its own main function.